requirements.txt
# Telegram Bot
python-telegram-bot==21.3

# HTTP requests
aiohttp==3.9.5
requests==2.32.3

# Database
aiosqlite==0.20.0

# Environment variables
python-dotenv==1.0.1

# Date/time handling
pytz==2024.1

# Scheduling
APScheduler==3.10.4

# JSON Schema validation
pydantic==2.7.4

# TOML parsing
toml==0.10.2

# Logging
structlog==24.2.0
